reorganizing monitor.js and updateLogicExecution.js
